Adding all endpoints to mta.js
