+++ src/mta.js
@@ -8,11 +8,41 @@ const request = require('request');
  * @returns {array} timeArray -- returns array with an array of JSON objects. JSON objects contain time in UNIX timestamp
  * @author Daniel Yoo (github: DanielY-Yoo)
  */
+
+const KEY = "c5881ea19e9d80654a41b10353585ef0";
+
+const dataFeeds = {
+  '1': `http://datamine.mta.info/mta_esi.php?key=${KEY}&feed_id=1`,
+  '2': `http://datamine.mta.info/mta_esi.php?key=${KEY}&feed_id=1`,
+  '3': `http://datamine.mta.info/mta_esi.php?key=${KEY}&feed_id=1`,
+  '4': `http://datamine.mta.info/mta_esi.php?key=${KEY}&feed_id=1`,
+  '5': `http://datamine.mta.info/mta_esi.php?key=${KEY}&feed_id=1`,
+  '6': `http://datamine.mta.info/mta_esi.php?key=${KEY}&feed_id=1`,
+  'S': `http://datamine.mta.info/mta_esi.php?key=${KEY}&feed_id=1`,
+  'A': `http://datamine.mta.info/mta_esi.php?key=${KEY}&feed_id=26`,
+  'C': `http://datamine.mta.info/mta_esi.php?key=${KEY}&feed_id=26`,
+  'E': `http://datamine.mta.info/mta_esi.php?key=${KEY}&feed_id=26`,
+  'S': `http://datamine.mta.info/mta_esi.php?key=${KEY}&feed_id=26`,
+  'N': `http://datamine.mta.info/mta_esi.php?key=${KEY}&feed_id=16`,
+  'Q': `http://datamine.mta.info/mta_esi.php?key=${KEY}&feed_id=16`,
+  'R': `http://datamine.mta.info/mta_esi.php?key=${KEY}&feed_id=16`,
+  'W': `http://datamine.mta.info/mta_esi.php?key=${KEY}&feed_id=16`,
+  'B': `http://datamine.mta.info/mta_esi.php?key=${KEY}&feed_id=21`,
+  'D': `http://datamine.mta.info/mta_esi.php?key=${KEY}&feed_id=21`,
+  'F': `http://datamine.mta.info/mta_esi.php?key=${KEY}&feed_id=21`,
+  'M': `http://datamine.mta.info/mta_esi.php?key=${KEY}&feed_id=21`,
+  'L': `http://datamine.mta.info/mta_esi.php?key=${KEY}&feed_id=2`,
+  'G': `http://datamine.mta.info/mta_esi.php?key=${KEY}&feed_id=31`,
+  'J': `http://datamine.mta.info/mta_esi.php?key=${KEY}&feed_id=36`,
+  'Z': `http://datamine.mta.info/mta_esi.php?key=${KEY}&feed_id=36`,
+  '7': `http://datamine.mta.info/mta_esi.php?key=${KEY}&feed_id=51`,
+}
+
 function getArrivalTimeList(trainType) {
 
     const requestSettings = {
         method: 'GET',
-        url: 'http://datamine.mta.info/mta_esi.php?key=4a9f2f0d04ff986286f7739acb65054d&feed_id=1',
+        url: dataFeeds[trainType],
         encoding: null
     };
 
@@ -64,6 +94,12 @@ function getArrivalTimeList(trainType) {
     });
 }
 
-getArrivalTimeList("6").then((timeArray) => {
-    console.log(timeArray);
+getArrivalTimeList('6').then((timeArray) => {
+    console.log(timeArray[0]);
+});
+getArrivalTimeList('C').then((timeArray) => {
+    console.log(timeArray[0]);
+});
+getArrivalTimeList('J').then((timeArray) => {
+    console.log(timeArray[0]);
 });
